migrate native components + note Mint/Transfer as deprecated
